+++ src/classes/constants.class.js
@@ -67,7 +67,11 @@ export class Constants {
                 changes: {
                     enabled: true,
                     interval: 60000,            // 60*1000=1mn en ms
-                    since: 600000               // 10*60*1000=10mn en ms: when starting, rewind 10mn to init the metrics
+                    since: 600000,              // 10*60*1000=10mn en ms: when starting, rewind 10mn to init the metrics
+                    traces: {
+                        rpc: false,
+                        metrics: false
+                    }
                 }
             },
             interaction: {